Creation des objets vide
