+++ app/src/main/java/fr/ecole/eni/lokacar/bean/ChiffreAffaire.java
@@ -1,0 +1,4 @@
+package fr.ecole.eni.lokacar.bean;
+
+public class ChiffreAffaire {
+}
