--- app/src/test/java/bean/Agence.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Agence {
-}
