--- app/src/test/java/bean/ChiffreAffaire.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class ChiffreAffaire {
-}
