--- app/src/test/java/bean/Client.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Client {
-}
