--- app/src/test/java/bean/DetailsModel.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class DetailsModel {
-}
