--- app/src/test/java/bean/Location.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Location {
-}
