--- app/src/test/java/bean/Marque.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Marque {
-}
