--- app/src/test/java/bean/Model.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Model {
-}
