--- app/src/test/java/bean/Salarie.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Salarie {
-}
