--- app/src/test/java/bean/Voiture.java
@@ -1,4 +1,0 @@
-package bean;
-
-public class Voiture {
-}
